Null-Check on Class from ServiceInfo (clz!=null)
Class obtained from ServiceInfo can be null. For example while working with Amazon Device Messaging you cannot export a jar with this class (because it won't work on Kindle Fire then). While launching on non-Kindle devide this class is not really used (and application would work fine) but Mechanoid will throw NPE due to lack of this service Class.